server/requirements.txt
### Librerias actualizadas con Flask 2.2.2 y Python 3.11.0 ###
aniso8601==9.0.1
autopep8==2.0.0
Cerberus==1.3.4
click==8.1.3
colorama==0.4.6
distlib==0.3.6
dnspython==2.3.0
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
jsons==1.6.3
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
packaging==23.0
platformdirs==2.6.2
protobuf==3.20.3
pycodestyle==2.9.1
pymongo==4.3.3
PyMySQL==1.0.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
tomli==2.0.1
typish==1.9.3
virtualenv==20.17.1
Werkzeug==2.2.2